+++ Modules/Intent.Modules.AzureFunctions.Dispatch.MediatR/FactoryExtensions/AzureFunctionsMediatRDispatchInstaller.cs
@@ -1,6 +1,3 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using Intent.AzureFunctions.Api;
 using Intent.Engine;
 using Intent.Modelers.Services.Api;
@@ -17,6 +14,10 @@ using Intent.Modules.Common.TypeResolution;
 using Intent.Modules.Metadata.WebApi.Models;
 using Intent.Plugins.FactoryExtensions;
 using Intent.RoslynWeaver.Attributes;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Net;
 
 [assembly: DefaultIntentManaged(Mode.Fully)]
 [assembly: IntentTemplate("Intent.ModuleBuilder.Templates.FactoryExtension", Version = "1.0")]
@@ -110,7 +111,12 @@ namespace Intent.Modules.AzureFunctions.Dispatch.MediatR.FactoryExtensions
 
         private static CSharpStatement GetReturnStatement(IAzureFunctionModel operationModel)
         {
-            var endpoint = HttpEndpointModelFactory.GetEndpoint(operationModel.InternalElement, string.Empty);
+            IHttpEndpointModel? endpoint = null;
+            if (operationModel.TriggerType == TriggerType.HttpTrigger)
+            {
+                endpoint = HttpEndpointModelFactory.GetEndpoint(operationModel.InternalElement, string.Empty);
+            }
+            
             switch (endpoint?.Verb)
             {
                 case HttpVerb.Get: