+++ Modules/Intent.Modules.AzureFunctions.OpenApi/NugetPackages.cs
@@ -19,23 +19,23 @@ namespace Intent.Modules.AzureFunctions.OpenApi
             NugetRegistry.Register(MicrosoftAzureFunctionsWorkerExtensionsOpenApiPackageName,
                 (framework) => (framework.Major, framework.Minor) switch
                     {
-                        ( >= 7, >= 0) => new PackageVersion("1.5.1")
+                        ( >= 7, >= 0) => new PackageVersion("1.6.0")
                             .WithNugetDependency("Microsoft.Azure.Core.NewtonsoftJson", "1.0.0")
                             .WithNugetDependency("Microsoft.Azure.Functions.Worker.Core", "1.8.0")
                             .WithNugetDependency("Microsoft.Azure.Functions.Worker.Extensions.Http", "3.0.13")
-                            .WithNugetDependency("Microsoft.Azure.WebJobs.Extensions.OpenApi.Core", "1.5.1")
+                            .WithNugetDependency("Microsoft.Azure.WebJobs.Extensions.OpenApi.Core", "1.6.0")
                             .WithNugetDependency("YamlDotNet", "12.0.1"),
-                        ( >= 6, >= 0) => new PackageVersion("1.5.1")
+                        ( >= 6, >= 0) => new PackageVersion("1.6.0")
                             .WithNugetDependency("Microsoft.Azure.Core.NewtonsoftJson", "1.0.0")
                             .WithNugetDependency("Microsoft.Azure.Functions.Worker.Core", "1.8.0")
                             .WithNugetDependency("Microsoft.Azure.Functions.Worker.Extensions.Http", "3.0.13")
-                            .WithNugetDependency("Microsoft.Azure.WebJobs.Extensions.OpenApi.Core", "1.5.1")
+                            .WithNugetDependency("Microsoft.Azure.WebJobs.Extensions.OpenApi.Core", "1.6.0")
                             .WithNugetDependency("YamlDotNet", "12.0.1"),
-                        ( >= 2, >= 0) => new PackageVersion("1.5.1")
+                        ( >= 2, >= 0) => new PackageVersion("1.6.0")
                             .WithNugetDependency("Microsoft.Azure.Core.NewtonsoftJson", "1.0.0")
                             .WithNugetDependency("Microsoft.Azure.Functions.Worker.Core", "1.8.0")
                             .WithNugetDependency("Microsoft.Azure.Functions.Worker.Extensions.Http", "3.0.13")
-                            .WithNugetDependency("Microsoft.Azure.WebJobs.Extensions.OpenApi.Core", "1.5.1")
+                            .WithNugetDependency("Microsoft.Azure.WebJobs.Extensions.OpenApi.Core", "1.6.0")
                             .WithNugetDependency("YamlDotNet", "12.0.1"),
                         _ => throw new Exception($"Unsupported Framework `{framework.Major}` for NuGet package '{MicrosoftAzureFunctionsWorkerExtensionsOpenApiPackageName}'"),
                     }
@@ -43,7 +43,10 @@ namespace Intent.Modules.AzureFunctions.OpenApi
             NugetRegistry.Register(MicrosoftAzureWebJobsExtensionsOpenApiPackageName,
                 (framework) => (framework.Major, framework.Minor) switch
                     {
-                        ( >= 2, >= 0) => new PackageVersion("1.5.1"),
+                        ( >= 2, >= 0) => new PackageVersion("1.6.0")
+                            .WithNugetDependency("Microsoft.Azure.WebJobs.Extensions.OpenApi.Core", "1.6.0")
+                            .WithNugetDependency("Microsoft.Azure.WebJobs.Script.Abstractions", "1.0.0-preview")
+                            .WithNugetDependency("YamlDotNet", "12.0.1"),
                         _ => throw new Exception($"Unsupported Framework `{framework.Major}` for NuGet package '{MicrosoftAzureWebJobsExtensionsOpenApiPackageName}'"),
                     }
                 );