+++ Tests/AzureFunctions.MongoDb/AzureFunctions.MongoDb.Api/Configuration/AzureKeyVaultConfiguration.cs
@@ -1,0 +1,49 @@
+using System;
+using Azure.Core;
+using Azure.Identity;
+using Intent.RoslynWeaver.Attributes;
+using Microsoft.Extensions.Configuration;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.Azure.KeyVault.AzureKeyVaultConfiguration", Version = "1.0")]
+
+namespace AzureFunctions.MongoDb.Api.Configuration
+{
+    public static class AzureKeyVaultConfiguration
+    {
+        public static void ConfigureAzureKeyVault(this IConfigurationBuilder builder, IConfiguration configuration)
+        {
+            var credential = GetTokenCredential(configuration);
+            var endpoint = configuration["KeyVault:Endpoint"];
+
+            if (string.IsNullOrWhiteSpace(endpoint))
+            {
+                throw new InvalidOperationException("Configuration 'KeyVault:Endpoint' is not set");
+            }
+            builder.AddAzureKeyVault(new Uri(endpoint), credential);
+        }
+
+        private static TokenCredential GetTokenCredential(IConfiguration configuration)
+        {
+            if (!string.IsNullOrWhiteSpace(configuration["KeyVault:TenantId"]) &&
+                !string.IsNullOrWhiteSpace(configuration["KeyVault:ClientId"]) &&
+                !string.IsNullOrWhiteSpace(configuration["KeyVault:Secret"]))
+            {
+                // Manually specify the connection details for Azure Key Vault.
+                // Its recommended to store the 'Secret' inside the .NET User Secret's secrets.json file.
+                return new ClientSecretCredential(configuration["KeyVault:TenantId"], configuration["KeyVault:ClientId"], configuration["KeyVault:Secret"]);
+            }
+
+            if (!string.IsNullOrWhiteSpace(configuration["KeyVault:ClientId"]))
+            {
+                // Connect to Azure Key Vault using the configured App Client Id.
+                return new DefaultAzureCredential(new DefaultAzureCredentialOptions
+                {
+                    ManagedIdentityClientId = configuration["KeyVault:ClientId"]
+                });
+            }
+            // Use the default discovery mechanisms to connect to Azure Key Vault.
+            return new DefaultAzureCredential();
+        }
+    }
+}